backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.0
pydantic==2.8.0
openai==1.40.0
python-dotenv==1.0.1
httpx==0.27.0
pymongo==4.8.0
